Move balance functions into ui-logic package
