+++ packages/bierzo-wallet/src/components/Header/index.stories.tsx
@@ -4,11 +4,11 @@ import { Block, Hairline, Typography } from "medulas-react-components";
 import * as React from "react";
 import { ReadonlyDate } from "readonly-date";
 import { DeepPartial } from "redux";
+import { stringToAmount } from "ui-logic";
 
 import { BwUnknownProps } from "../../logic/transactions/types/BwUnkownTransaction";
 import { ProcessedSendTransaction } from "../../store/notifications";
 import { RootState } from "../../store/reducers";
-import { stringToAmount } from "../../utils/balances";
 import DecoratedStorybook, { WALLET_ROOT } from "../../utils/storybook";
 import Header from "./index";
