+++ packages/ui-logic/types/index.d.ts
@@ -1,3 +1,4 @@
+export { amountToGwei, amountToNumber, amountToString, stringToAmount } from "./balances";
 export { singleton } from "./singleton";
 export { ellipsify } from "./strings";
 export { whenTrue } from "./testing";
